requirements.txt
streamlit
scikit-learn
pandas
numpyS